[Data Streams] Fix forward-compatibility test failures (#223129)
Fixes https://github.com/elastic/kibana/issues/221224

## Summary

This PR fixes the forward-compatibility test for index mode, which
failed because the index mode was added to Get Data Streams API in 8.19
and 9.1 (see https://github.com/elastic/elasticsearch/pull/122486), so
if Kibana 8.19 is run with Es 9.0, the index mode is always displayed as
"Standard" because Es doesn't return an index mode field.

In this PR, we separate all index mode-related tests into a separate
file, and only run it for Es versions 8.19 or 9.1+.

(cherry picked from commit d12b19608af35df7ad1c3b38371819b90c926676)
